+++ packages/entities/entities-vaults/sandbox/pages/VaultFormPage.vue
@@ -58,6 +58,7 @@ const kongManagerConfig = ref<KongManagerVaultFormConfig>({
   conjurVaultProviderAvailable: true,
   base64FieldAvailable: true,
   hcvCertMethodAvailable: true,
+  hcvOauth2MethodAvailable: true,
 })
 
 const onError = (error: AxiosError) => {